Cleanup existing gateway if launching the boat: fixes issue#11.
